Pipeline visualisations

A Pipeline object is automatically constructed as calls to make_*() are made. It stores the relationships between targets, dependencies, and sources.

See also

Public fields

edges

A data frame

nodes

A data frame

Methods

Public methods

Pipeline$add_source_segment()

Pipeline$add_recipe_segment()

Pipeline$style_nodes()

Pipeline$print()

Pipeline$save()

Pipeline$clone()

Method add_source_segment()

Add a pipeline segment corresponding to a make_with_source() call

Usage

Pipeline$add_source_segment(source, targets, dependencies)

Arguments

source

The path to an R script which makes the targets

targets

A character vector of paths to files

dependencies

A character vector of paths to files which the targets depend on

Returns

self

Method add_recipe_segment()

Add a pipeline segment corresponding to a make_with_recipe() call

Usage

Pipeline$add_recipe_segment(recipe, targets, dependencies)

Arguments

recipe

A character vector containing a deparsed expression, which would make the targets if evaluated.

targets

A character vector of paths to files

dependencies

A character vector of paths to files which the targets depend on

Returns

self

Method style_nodes()

Style pipeline nodes

Usage

Pipeline$style_nodes()

Returns

self

Method print()

Display pipeline

Usage

Pipeline$print()

Returns

self

Method save()

Save pipeline

Usage

Pipeline$save(file, ...)

Arguments

file

File to save HTML into

...

Additional arguments to pass to visNetwork::visSave()

Returns

self

Method clone()

The objects of this class are cloneable with this method.

Usage

Pipeline$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.