Make targets out of dependencies using a recipe

Make targets out of dependencies using a recipe

make_with_recipe(
  recipe,
  targets,
  dependencies,
  envir = parent.frame(),
  quiet = getOption("piper.quiet"),
  ...
)

Arguments

recipe
A chunk of R code which makes the targets

targets
A character vector of paths to files

dependencies
A character vector of paths to files which the targets depend on

envir
The environment in which to run recipe

quiet
A logical determining whether or not messages are signaled

...
Additional parameters to pass to base::eval()

Value

The result of evaluating the recipe if the targets are out of date otherwise `NULL``

See also

Examples

if (FALSE) {
# Merge files in fresh environment if raw data has been updated
# since last merged
make_with_recipe(
  recipe = {
    dat <- readRDS("data/raw_data.Rds")
    pop <- readRDS("data/pop_data.Rds")
    merged_dat <- merge(dat, pop, by = "id")
    saveRDS(merged_dat, "data/merged_data.Rds")
  },
  targets = "data/merged_data.Rds",
  dependencies = c("data/raw_data.Rds", "data/raw_pop.Rds"),
  envir = new.env()
)
}