Access and interface with Pipeline.

get_pipeline(), set_pipeline() access and modify the current active pipeline, while all other helper functions do not affect the active pipeline

is_pipeline(pipeline)

set_pipeline(pipeline)

get_pipeline()

Arguments

pipeline
A pipeline. See Pipeline for more details.

See also

Examples

if (FALSE) {
 # Build up a pipeline from scratch and save it out
 set_pipeline(Pipeline$new())
 # A series of `make_with_*()` blocks go here...
 saveRDS(get_pipeline(), "data/my_pipeline.Rds")
}